Fix blended color conversions
prompt: fix explicit color component conversions for the custom build

+++ src/SumatraPDF.cpp
@@ -3089,9 +3089,9 @@ static void UpdateToolbarSidebarText(MainWindow* win) {
 
 static Color BlendColor(Color base, Color overlay, int overlayPercent) {
     int basePercent = 100 - overlayPercent;
-    return MkRgb((GetRed(base) * basePercent + GetRed(overlay) * overlayPercent) / 100,
-                 (GetGreen(base) * basePercent + GetGreen(overlay) * overlayPercent) / 100,
-                 (GetBlue(base) * basePercent + GetBlue(overlay) * overlayPercent) / 100);
+    return MkRgb(u8((GetRed(base) * basePercent + GetRed(overlay) * overlayPercent) / 100),
+                 u8((GetGreen(base) * basePercent + GetGreen(overlay) * overlayPercent) / 100),
+                 u8((GetBlue(base) * basePercent + GetBlue(overlay) * overlayPercent) / 100));
 }
 
 static Color DwmFrameBorderColorForCurrentTheme(MainWindow* win) {
